added

+++ app/Repositories/ProfileRepository.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Repositories;
+
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
@@ -9,40 +10,33 @@ class ProfileRepository implements ProfileRepositoryInterface
     public function getUser()
     {
         $user = Auth::user();
-
         $user->load([
             'posts' => function ($query) {
-                $query->with(['likes', 'comments.user'])
-                    ->latest();
+                $query->with(['likes', 'comments.user'])->latest();
             }
         ]);
-
         return $user;
     }
 
     public function updateProfile(array $data)
     {
-        // Logged-in user
         $user = Auth::user();
 
-        // Update basic fields
         $user->fill([
             'name' => $data['name'],
             'email' => $data['email'],
         ]);
 
-        // Delete old profile image if exists
+        // Profile image update
         if (!empty($data['profile_image'])) {
-            Storage::disk('public')->delete($user->profile_image);
-        }
+            if ($user->profile_image) {
+                Storage::disk('public')->delete($user->profile_image);
+            }
 
-        // Store new profile image
-        if (!empty($data['profile_image'])) {
             $user->profile_image = $data['profile_image']->store('profile_images', 'public');
         }
-        // Save  to database
-        $user->save();
 
+        $user->save();
         return $user->fresh();
     }
 
@@ -50,43 +44,36 @@ class ProfileRepository implements ProfileRepositoryInterface
     {
         $user = Auth::user();
 
-        // Delete old cover image if exists
-        if ($user->cover_image) {
-            Storage::disk('public')->delete($user->cover_image);
+        if (!empty($data['cover_image'])) {
+            if ($user->cover_image) {
+                Storage::disk('public')->delete($user->cover_image);
+            }
+
+            $user->cover_image = $data['cover_image']->store('cover_images', 'public');
         }
 
-        // Store new cover image
-        if (isset($data['userCover'])) {
-            $user->cover_image = $data['userCover']->store('cover_images', 'public');
-
-        }
         $user->save();
-        return $user;
+
+        return $user->fresh();
     }
 
     public function updateProfileImage(array $data)
     {
-
         $user = Auth::user();
 
-        //old image delete
+        // Profile image update
         if (!empty($data['profile_image'])) {
-            Storage::disk('public')->delete($user->profile_image ?? '');
+            if ($user->profile_image) {
+                Storage::disk('public')->delete($user->profile_image);
+            }
 
-            // new Image Save
             $user->profile_image = $data['profile_image']->store('profile_images', 'public');
-            $user->save();
         }
 
-        return $user;
+        $user->save();
 
+        // dd($user->profile_image); 
+
+        return $user;
     }
 }
-
-
-
-
-
-
-
-
